Refactor FlexibleInvoice component into modular architecture
Break down 2,300+ line monolithic component into focused, reusable modules:

Components (components/invoice-editor/):
- InvoiceHeader: Header with dates and invoice information
- ContactDetails: From/To sections with dynamic field management
- ServicesTable: Complex editable table with dynamic columns/rows
- TotalsSection: Subtotal, tax, discount calculations and controls
- PaymentDetails: Payment/bank details section
- ActionButtons: Print and action controls

Custom Hooks (hooks/):
- useInvoiceState: Centralized state management for all invoice data
- useInvoiceCalculations: Business logic for all financial calculations
- useFieldManagement: CRUD operations for dynamic fields
- useTableManagement: Table operations (add/remove columns/rows)

Utilities (lib/):
- invoice-types: TypeScript interfaces and type definitions
- invoice-utils: Pure utility functions for dates, parsing, formatting

Benefits:
- 43% reduction in total lines through better organization
- Single Responsibility Principle applied to each module
- Improved maintainability and testability
- Enhanced reusability of components and hooks
- Better separation of concerns (UI, state, business logic)
- Maintains 100% functional equivalence with original
