+++ components/FlexibleInvoice.tsx
@@ -1,289 +1,78 @@
 'use client'
 
-import React, { useState, useRef, useEffect, useCallback } from 'react'
+import React, { useRef, useEffect } from 'react'
 import CurrencySelector from './invoice-editor/CurrencySelector'
+import { InvoiceHeader } from './invoice-editor/InvoiceHeader'
+import { ContactDetails } from './invoice-editor/ContactDetails'
+import { ServicesTable } from './invoice-editor/ServicesTable'
+import { TotalsSection } from './invoice-editor/TotalsSection'
+import { PaymentDetails } from './invoice-editor/PaymentDetails'
+import { ActionButtons } from './invoice-editor/ActionButtons'
+import { useInvoiceState } from '../hooks/useInvoiceState'
+import { useInvoiceCalculations } from '../hooks/useInvoiceCalculations'
+import { useTableManagement } from '../hooks/useTableManagement'
 
 export default function FlexibleInvoice() {
   const invoiceRef = useRef<HTMLDivElement>(null)
   
-  // State for individual fields
-  const [fromFields, setFromFields] = useState([
-    { id: '1', label: 'Your Business Name', value: '', placeholder: 'Your Business Name' },
-    { id: '2', label: 'Street Address', value: '', placeholder: 'Street Address' },
-    { id: '3', label: 'City', value: '', placeholder: 'City' },
-    { id: '4', label: 'State/Province', value: '', placeholder: 'State/Province' },
-    { id: '5', label: 'Country', value: '', placeholder: 'Country' },
-    { id: '6', label: 'Email', value: '', placeholder: 'your.email@business.com' },
-    { id: '7', label: 'Phone', value: '', placeholder: '+1 (555) 123-4567' }
-  ])
-  
-  const [toFields, setToFields] = useState([
-    { id: '1', label: 'Client Name', value: '', placeholder: 'Client Name' },
-    { id: '2', label: 'Client Address', value: '', placeholder: 'Client Address' },
-    { id: '3', label: 'City', value: '', placeholder: 'City' },
-    { id: '4', label: 'State/Province', value: '', placeholder: 'State/Province' },
-    { id: '5', label: 'Country', value: '', placeholder: 'Country' },
-    { id: '6', label: 'Email', value: '', placeholder: 'client@email.com' }
-  ])
-  
-  const [paymentFields, setPaymentFields] = useState([
-    { id: '1', label: 'Account Name', value: '', placeholder: 'Your Full Name' },
-    { id: '2', label: 'Bank', value: '', placeholder: 'Your Bank Name' },
-    { id: '3', label: 'Account Number', value: '', placeholder: '0000 0000 00' },
-    { id: '4', label: 'Branch Code', value: '', placeholder: '000000' },
-    { id: '5', label: 'SWIFT Code', value: '', placeholder: 'BANKCODE' }
-  ])
-  
-  // Tax and discount state
-  const [taxEnabled, setTaxEnabled] = useState(false)
-  const [taxPercentage, setTaxPercentage] = useState('15')
-  const [discountEnabled, setDiscountEnabled] = useState(false)  
-  const [discountPercentage, setDiscountPercentage] = useState('10')
-  
-  // Currency state - default to ZAR to maintain current behavior
-  const [currencyCode, setCurrencyCode] = useState('ZAR')
-  const [currencySymbol, setCurrencySymbol] = useState('R')
+  // All state management
+  const {
+    fromFields,
+    setFromFields,
+    toFields,
+    setToFields,
+    paymentFields,
+    setPaymentFields,
+    taxEnabled,
+    setTaxEnabled,
+    taxPercentage,
+    setTaxPercentage,
+    discountEnabled,
+    setDiscountEnabled,
+    discountPercentage,
+    setDiscountPercentage,
+    currencyCode,
+    setCurrencyCode,
+    currencySymbol,
+    setCurrencySymbol,
+    invoiceDate,
+    setInvoiceDate,
+    dueDate,
+    setDueDate,
+    invoiceTitle,
+    setInvoiceTitle,
+    fromTitle,
+    setFromTitle,
+    toTitle,
+    setToTitle,
+    paymentTitle,
+    setPaymentTitle,
+    invoiceNumberLabel,
+    setInvoiceNumberLabel,
+    dateLabel,
+    setDateLabel,
+    dueDateLabel,
+    setDueDateLabel,
+    totalLabel,
+    setTotalLabel,
+    columns,
+    setColumns,
+    rows,
+    setRows
+  } = useInvoiceState()
 
-  // Date helper functions
-  const getCurrentDate = () => {
-    const today = new Date()
-    return today.toISOString().split('T')[0] // YYYY-MM-DD format
-  }
+  // Calculations
+  const { calculateRowAmount, totals } = useInvoiceCalculations({
+    columns,
+    rows,
+    taxEnabled,
+    taxPercentage,
+    discountEnabled,
+    discountPercentage
+  })
 
-  const getEndOfMonth = () => {
-    const today = new Date()
-    const lastDay = new Date(today.getFullYear(), today.getMonth() + 1, 0)
-    return lastDay.toISOString().split('T')[0] // YYYY-MM-DD format
-  }
-
-  // Date state
-  const [invoiceDate, setInvoiceDate] = useState(getCurrentDate())
-  const [dueDate, setDueDate] = useState(getEndOfMonth())
-  
-  // Section headers state
-  const [invoiceTitle, setInvoiceTitle] = useState('INVOICE')
-  const [fromTitle, setFromTitle] = useState('From:')
-  const [toTitle, setToTitle] = useState('To:')
-  const [paymentTitle, setPaymentTitle] = useState('Payment Details:')
-  
-  // Field labels state
-  const [invoiceNumberLabel, setInvoiceNumberLabel] = useState('Invoice #:')
-  const [dateLabel, setDateLabel] = useState('Issue Date:')
-  const [dueDateLabel, setDueDateLabel] = useState('Due Date:')
-  const [totalLabel, setTotalLabel] = useState('Total Amount Due:')
-
-  // Format date for display (DD/MM/YYYY)
-  const formatDateForDisplay = (dateString: string) => {
-    if (!dateString) return ''
-    const date = new Date(dateString)
-    const day = date.getDate().toString().padStart(2, '0')
-    const month = (date.getMonth() + 1).toString().padStart(2, '0')
-    const year = date.getFullYear()
-    return `${day}/${month}/${year}`
-  }
-  
-  // Table state - start with default columns and one row
-  const [columns, setColumns] = useState([
-    { id: 'description', name: 'Description', width: '50%', align: 'left', isDescription: true },
-    { id: '2', name: 'Rate', width: '20%', align: 'right', isDescription: false },
-    { id: '3', name: 'Quantity', width: '15%', align: 'center', isDescription: false },
-    { id: 'amount', name: 'Amount', width: '15%', align: 'right', isAmount: true }
-  ])
-  
-  const [rows, setRows] = useState([
-    { 
-      id: '1', 
-      cells: {
-        'description': { name: '', description: '' },
-        '2': '',
-        '3': '', 
-        'amount': ''
-      }
-    }
-  ])
-
-  // Helper function to parse numeric value from string
-  const parseNumericValue = (value: string): number => {
-    if (!value) return 0
-    // Remove currency symbols and spaces
-    const cleanValue = value.replace(/[^0-9.,-]/g, '')
-    const numValue = parseFloat(cleanValue.replace(',', ''))
-    return isNaN(numValue) ? 0 : numValue
-  }
-
-  // Calculate amount for a row based on numeric columns
-  const calculateRowAmount = useCallback((row: any): string => {
-    let result = 1
-    let hasValues = false
-    
-    columns.forEach(col => {
-      if (!col.isDescription && !col.isAmount && row.cells[col.id]) {
-        const numValue = parseNumericValue(String(row.cells[col.id]))
-        if (numValue !== 0) {
-          result *= numValue
-          hasValues = true
-        }
-      }
-    })
-    
-    return hasValues ? result.toFixed(2) : ''
-  }, [columns])
-
-  // Calculate subtotal from all amount values
-  const calculateSubtotal = (): string => {
-    let total = 0
-    rows.forEach(row => {
-      const amountCol = columns.find(col => col.isAmount)
-      if (amountCol && (row.cells as any)[amountCol.id]) {
-        total += parseNumericValue(String((row.cells as any)[amountCol.id]))
-      }
-    })
-    return total.toFixed(2)
-  }
-
-  // Calculate tax amount
-  const calculateTaxAmount = (): string => {
-    if (!taxEnabled) return '0.00'
-    const subtotal = parseFloat(calculateSubtotal())
-    const taxRate = parseNumericValue(taxPercentage) / 100
-    return (subtotal * taxRate).toFixed(2)
-  }
-
-  // Calculate discount amount
-  const calculateDiscountAmount = (): string => {
-    if (!discountEnabled) return '0.00'
-    const subtotal = parseFloat(calculateSubtotal())
-    const discountRate = parseNumericValue(discountPercentage) / 100
-    return (subtotal * discountRate).toFixed(2)
-  }
-
-  // Calculate final total
-  const calculateTotal = (): string => {
-    const subtotal = parseFloat(calculateSubtotal())
-    const tax = parseFloat(calculateTaxAmount())
-    const discount = parseFloat(calculateDiscountAmount())
-    return (subtotal + tax - discount).toFixed(2) // Discount is subtracted
-  }
-
-  // Helper function to get placeholder text for table cells
-  const getPlaceholderText = (column: any): string => {
-    const columnName = column.name.toLowerCase()
-    if (columnName.includes('rate') || columnName.includes('price')) {
-      return '50.00'
-    }
-    if (columnName.includes('quantity') || columnName.includes('qty')) {
-      return '1'
-    }
-    if (column.isAmount) {
-      return '50.00'
-    }
-    return 'Enter value'
-  }
-
-  // Helper function to render table cell content
-  const renderCellContent = (row: any, column: any) => {
-    if (column.isDescription && typeof row.cells[column.id] === 'object') {
-      // Special handling for description column
-      return (
-        <div style={{ display: 'flex', flexDirection: 'column', gap: '4px' }}>
-          <input
-            type="text"
-            value={row.cells[column.id].name}
-            placeholder="Item Name"
-            onChange={(e) => updateCell(row.id, column.id, {
-              ...row.cells[column.id],
-              name: e.target.value
-            })}
-            style={{
-              backgroundColor: '#ffffff',
-              border: '2px solid #e2e8f0',
-              borderRadius: '6px',
-              padding: '8px 12px',
-              fontSize: '14px',
-              outline: 'none',
-              fontWeight: '600',
-              width: '100%',
-              transition: 'all 0.2s ease',
-              color: '#1f2937'
-            }}
-            onFocus={(e) => {
-              e.target.style.borderColor = '#667eea'
-              e.target.style.boxShadow = '0 0 0 3px rgba(102, 126, 234, 0.1)'
-            }}
-            onBlur={(e) => {
-              e.target.style.borderColor = '#e2e8f0'
-              e.target.style.boxShadow = 'none'
-            }}
-          />
-          <input
-            type="text"
-            value={row.cells[column.id].description}
-            placeholder="Description of services provided"
-            onChange={(e) => updateCell(row.id, column.id, {
-              ...row.cells[column.id],
-              description: e.target.value
-            })}
-            style={{
-              backgroundColor: '#ffffff',
-              border: '2px solid #e2e8f0',
-              borderRadius: '6px',
-              padding: '8px 12px',
-              fontSize: '14px',
-              outline: 'none',
-              width: '100%',
-              transition: 'all 0.2s ease',
-              color: '#6b7280'
-            }}
-            onFocus={(e) => {
-              e.target.style.borderColor = '#667eea'
-              e.target.style.boxShadow = '0 0 0 3px rgba(102, 126, 234, 0.1)'
-            }}
-            onBlur={(e) => {
-              e.target.style.borderColor = '#e2e8f0'
-              e.target.style.boxShadow = 'none'
-            }}
-          />
-        </div>
-      )
-    }
-
-    // Regular cell - Amount is read-only calculated field, others are editable
-    return (
-      <input
-        type="text"
-        value={row.cells[column.id]}
-        placeholder={getPlaceholderText(column)}
-        onChange={(e) => updateCell(row.id, column.id, e.target.value)}
-        readOnly={column.isAmount}
-        style={{
-          backgroundColor: column.isAmount ? '#f8fafc' : '#ffffff',
-          border: column.isAmount ? '2px solid #e5e7eb' : '2px solid #e2e8f0',
-          borderRadius: '6px',
-          padding: '8px 12px',
-          fontSize: '14px',
-          outline: 'none',
-          width: '100%',
-          textAlign: column.align as 'left' | 'center' | 'right',
-          cursor: column.isAmount ? 'not-allowed' : 'text',
-          transition: 'all 0.2s ease',
-          fontWeight: column.isAmount ? '600' : '500',
-          color: column.isAmount ? '#374151' : '#1f2937'
-        }}
-        onFocus={(e) => {
-          if (!column.isAmount) {
-            e.target.style.borderColor = '#667eea'
-            e.target.style.boxShadow = '0 0 0 3px rgba(102, 126, 234, 0.1)'
-          }
-        }}
-        onBlur={(e) => {
-          if (!column.isAmount) {
-            e.target.style.borderColor = '#e2e8f0'
-            e.target.style.boxShadow = 'none'
-          }
-        }}
-      />
-    )
-  }
+  // Table management
+  const { updateCell } = useTableManagement()
 
   // Update amounts when numeric values change
   useEffect(() => {
@@ -312,158 +101,10 @@ export default function FlexibleInvoice() {
     if (hasChanges) {
       setRows(updatedRows)
     }
-  }, [rows, columns, calculateRowAmount])
+  }, [rows, columns, calculateRowAmount, setRows])
 
-  const addRow = () => {
-    const newRow = {
-      id: Date.now().toString(),
-      cells: columns.reduce((acc, col) => {
-        if (col.isDescription) {
-          acc[col.id] = { name: '', description: '' }
-        } else {
-          acc[col.id] = ''
-        }
-        return acc
-      }, {} as any)
-    }
-    setRows([...rows, newRow])
-  }
-
-  const removeRow = (rowId: string) => {
-    if (rows.length > 1) {
-      setRows(rows.filter(row => row.id !== rowId))
-    }
-  }
-
-  const addColumn = (afterColumnId?: string) => {
-    const newColumn = {
-      id: Date.now().toString(),
-      name: 'New Column',
-      width: '10%',
-      align: 'center' as const,
-      isDescription: false
-    }
-    
-    if (afterColumnId) {
-      // Insert after specific column
-      const afterIndex = columns.findIndex(col => col.id === afterColumnId)
-      const newColumns = [...columns]
-      newColumns.splice(afterIndex + 1, 0, newColumn)
-      setColumns(newColumns)
-    } else {
-      // Default: Insert before the Amount column if it exists, otherwise at the end
-      const amountColIndex = columns.findIndex(col => col.isAmount)
-      if (amountColIndex !== -1) {
-        const newColumns = [...columns]
-        newColumns.splice(amountColIndex, 0, newColumn)
-        setColumns(newColumns)
-      } else {
-        setColumns([...columns, newColumn])
-      }
-    }
-    
-    // Add empty cell to all existing rows
-    setRows(rows.map(row => ({
-      ...row,
-      cells: { ...row.cells, [newColumn.id]: '' }
-    })))
-  }
-
-  const removeColumn = (columnId: string) => {
-    const column = columns.find(col => col.id === columnId)
-    // Don't allow removing Description or Amount columns, or if only 2 columns left
-    if (column?.isDescription || column?.isAmount || columns.length <= 2) {
-      return
-    }
-    
-    setColumns(columns.filter(col => col.id !== columnId))
-    // Remove cells from all rows
-    setRows(rows.map(row => {
-      const newCells = { ...row.cells } as any
-      delete newCells[columnId]
-      return { ...row, cells: newCells }
-    }))
-  }
-
-  const updateColumnName = (columnId: string, newName: string) => {
-    setColumns(columns.map(col => 
-      col.id === columnId ? { ...col, name: newName } : col
-    ))
-  }
-
-  const updateCell = (rowId: string, columnId: string, value: any) => {
-    setRows(rows.map(row => 
-      row.id === rowId 
-        ? { ...row, cells: { ...row.cells, [columnId]: value } }
-        : row
-    ))
-  }
-
-  // Field management functions
-  const addFromField = () => {
-    const newField = {
-      id: Date.now().toString(),
-      label: 'New Field',
-      value: '',
-      placeholder: 'Enter value'
-    }
-    setFromFields([...fromFields, newField])
-  }
-
-  const removeFromField = (fieldId: string) => {
-    if (fromFields.length > 1) {
-      setFromFields(fromFields.filter(field => field.id !== fieldId))
-    }
-  }
-
-  const updateFromField = (fieldId: string, value: string) => {
-    setFromFields(fromFields.map(field => 
-      field.id === fieldId ? { ...field, value } : field
-    ))
-  }
-
-  const addToField = () => {
-    const newField = {
-      id: Date.now().toString(),
-      label: 'New Field',
-      value: '',
-      placeholder: 'Enter value'
-    }
-    setToFields([...toFields, newField])
-  }
-
-  const removeToField = (fieldId: string) => {
-    if (toFields.length > 1) {
-      setToFields(toFields.filter(field => field.id !== fieldId))
-    }
-  }
-
-  const updateToField = (fieldId: string, value: string) => {
-    setToFields(toFields.map(field => 
-      field.id === fieldId ? { ...field, value } : field
-    ))
-  }
-
-  const addPaymentField = () => {
-    const newField = {
-      id: Date.now().toString(),
-      label: 'New Field',
-      value: '',
-      placeholder: 'Enter value'
-    }
-    setPaymentFields([...paymentFields, newField])
-  }
-
-  const removePaymentField = (fieldId: string) => {
-    if (paymentFields.length > 1) {
-      setPaymentFields(paymentFields.filter(field => field.id !== fieldId))
-    }
-  }
-
-  const updatePaymentField = (fieldId: string, value: string) => {
-    setPaymentFields(paymentFields.map(field => 
-      field.id === fieldId ? { ...field, value } : field
-    ))
+  const handleCellUpdate = (rowId: string, columnId: string, value: any) => {
+    updateCell(rowId, columnId, value, rows, setRows)
   }
 
   return (
@@ -474,7 +115,6 @@ export default function FlexibleInvoice() {
       background: 'linear-gradient(135deg, #667eea 0%, #764ba2 100%)',
       padding: '40px 20px'
     }}>
-
       <div ref={invoiceRef} className="invoice-container" style={{
         maxWidth: '900px',
         margin: '0 auto',
@@ -484,1488 +124,101 @@ export default function FlexibleInvoice() {
         overflow: 'hidden'
       }}>
         <div style={{ padding: '40px' }}>
-        
-        {/* Invoice Header */}
-        <div className="header" style={{
-          display: 'flex',
-          justifyContent: 'space-between',
-          alignItems: 'flex-start',
-          marginBottom: '40px',
-          paddingBottom: '30px',
-          borderBottom: '3px solid #f1f3f5'
-        }}>
-          <div>
-            <h1 className="invoice-title" style={{
-              fontSize: '36px',
-              fontWeight: '800',
-              background: 'linear-gradient(135deg, #667eea 0%, #764ba2 100%)',
-              WebkitBackgroundClip: 'text',
-              WebkitTextFillColor: 'transparent',
-              backgroundClip: 'text',
-              margin: 0,
-              letterSpacing: '-0.02em',
-              minWidth: '200px',
-              display: 'inline-block',
-              position: 'relative'
-            }}>
-              <span
-                role="textbox"
-                aria-label="Invoice title"
-                tabIndex={0}
-                contentEditable
-                suppressContentEditableWarning
-                style={{
-                  outline: 'none',
-                  cursor: 'text',
-                  display: 'inline-block',
-                  padding: '8px 12px',
-                  borderRadius: '8px',
-                  transition: 'all 0.2s ease',
-                  border: '2px solid #e2e8f0',
-                  backgroundColor: '#f8fafc',
-                  background: 'linear-gradient(135deg, #667eea 0%, #764ba2 100%)',
-                  WebkitBackgroundClip: 'text',
-                  WebkitTextFillColor: 'transparent',
-                  backgroundClip: 'text'
-                }}
-                onFocus={(e) => {
-                  e.target.style.backgroundColor = '#ffffff'
-                  e.target.style.borderColor = '#667eea'
-                  e.target.style.boxShadow = '0 0 0 3px rgba(102, 126, 234, 0.1)'
-                }}
-                onBlur={(e) => {
-                  setInvoiceTitle(e.target.textContent || 'INVOICE')
-                  e.target.style.backgroundColor = '#f8fafc'
-                  e.target.style.borderColor = '#e2e8f0'
-                  e.target.style.boxShadow = 'none'
-                }}
-                onKeyDown={(e) => {
-                  if (e.key === 'Enter') {
-                    e.preventDefault()
-                    e.currentTarget.blur()
-                  }
-                }}
-              >
-                {invoiceTitle}
-              </span>
-            </h1>
-          </div>
-          <div className="invoice-details" style={{
-            textAlign: 'right'
-          }}>
-            <div style={{ 
-              marginBottom: '16px',
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'flex-end',
-              gap: '12px'
-            }}>
-              <span
-                role="textbox"
-                aria-label="Invoice number label"
-                tabIndex={0}
-                contentEditable
-                suppressContentEditableWarning
-                style={{
-                  fontSize: '15px',
-                  fontWeight: '600',
-                  color: '#374151',
-                  outline: 'none',
-                  cursor: 'text',
-                  padding: '6px 10px',
-                  borderRadius: '8px',
-                  transition: 'all 0.2s ease',
-                  backgroundColor: '#f8fafc',
-                  border: '2px solid #e2e8f0',
-                  minWidth: '110px',
-                  width: '110px',
-                  display: 'inline-block',
-                  textAlign: 'right',
-                  whiteSpace: 'nowrap'
-                }}
-                onFocus={(e) => {
-                  e.target.style.backgroundColor = '#ffffff'
-                  e.target.style.borderColor = '#667eea'
-                  e.target.style.boxShadow = '0 0 0 3px rgba(102, 126, 234, 0.1)'
-                }}
-                onBlur={(e) => {
-                  setInvoiceNumberLabel(e.target.textContent || 'Invoice #:')
-                  e.target.style.backgroundColor = '#f8fafc'
-                  e.target.style.borderColor = '#e2e8f0'
-                  e.target.style.boxShadow = 'none'
-                }}
-                onKeyDown={(e) => {
-                  if (e.key === 'Enter') {
-                    e.preventDefault()
-                    e.currentTarget.blur()
-                  }
-                }}
-              >
-                {invoiceNumberLabel}
-              </span>
-              <input
-                type="text"
-                placeholder="INV-2025-08"
-                style={{
-                  backgroundColor: '#f8fafc',
-                  padding: '10px 16px',
-                  borderRadius: '8px',
-                  border: '2px solid #e2e8f0',
-                  minWidth: '130px',
-                  width: '130px',
-                  fontSize: '14px',
-                  fontWeight: '500',
-                  outline: 'none',
-                  textAlign: 'right',
-                  transition: 'all 0.2s ease',
-                  color: '#1f2937'
-                }}
-                onFocus={(e) => {
-                  e.target.style.borderColor = '#667eea'
-                  e.target.style.backgroundColor = '#ffffff'
-                  e.target.style.boxShadow = '0 0 0 3px rgba(102, 126, 234, 0.1)'
-                }}
-                onBlur={(e) => {
-                  e.target.style.borderColor = '#e2e8f0'
-                  e.target.style.backgroundColor = '#f8fafc'
-                  e.target.style.boxShadow = 'none'
-                }}
-              />
-            </div>
-            <div style={{ 
-              marginBottom: '16px',
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'flex-end',
-              gap: '12px'
-            }}>
-              <span
-                role="textbox"
-                aria-label="Date label"
-                tabIndex={0}
-                contentEditable
-                suppressContentEditableWarning
-                style={{
-                  fontSize: '15px',
-                  fontWeight: '600',
-                  color: '#374151',
-                  outline: 'none',
-                  cursor: 'text',
-                  padding: '6px 10px',
-                  borderRadius: '8px',
-                  transition: 'all 0.2s ease',
-                  backgroundColor: '#f8fafc',
-                  border: '2px solid #e2e8f0',
-                  minWidth: '110px',
-                  width: '110px',
-                  display: 'inline-block',
-                  textAlign: 'right',
-                  whiteSpace: 'nowrap'
-                }}
-                onFocus={(e) => {
-                  e.target.style.backgroundColor = '#ffffff'
-                  e.target.style.borderColor = '#667eea'
-                  e.target.style.boxShadow = '0 0 0 3px rgba(102, 126, 234, 0.1)'
-                }}
-                onBlur={(e) => {
-                  setDateLabel(e.target.textContent || 'Issue Date:')
-                  e.target.style.backgroundColor = '#f8fafc'
-                  e.target.style.borderColor = '#e2e8f0'
-                  e.target.style.boxShadow = 'none'
-                }}
-                onKeyDown={(e) => {
-                  if (e.key === 'Enter') {
-                    e.preventDefault()
-                    e.currentTarget.blur()
-                  }
-                }}
-              >
-                {dateLabel}
-              </span> 
-              <div style={{ position: 'relative', display: 'inline-block' }}>
-                <input
-                  type="date"
-                  value={invoiceDate}
-                  onChange={(e) => setInvoiceDate(e.target.value)}
-                  style={{
-                    backgroundColor: '#f8fafc',
-                    padding: '10px 16px',
-                    borderRadius: '8px',
-                    border: '2px solid #e2e8f0',
-                    minWidth: '130px',
-                  width: '130px',
-                    fontSize: '14px',
-                    fontWeight: '500',
-                    outline: 'none',
-                    textAlign: 'right',
-                    cursor: 'pointer',
-                    transition: 'all 0.2s ease',
-                    color: '#1f2937'
-                  }}
-                  onFocus={(e) => {
-                    e.target.style.borderColor = '#667eea'
-                    e.target.style.backgroundColor = '#ffffff'
-                    e.target.style.boxShadow = '0 0 0 3px rgba(102, 126, 234, 0.1)'
-                  }}
-                  onBlur={(e) => {
-                    e.target.style.borderColor = '#e2e8f0'
-                    e.target.style.backgroundColor = '#f8fafc'
-                    e.target.style.boxShadow = 'none'
-                  }}
-                />
-                <span 
-                  style={{
-                    position: 'absolute',
-                    right: '16px',
-                    top: '10px',
-                    pointerEvents: 'none',
-                    fontSize: '14px',
-                    color: '#1f2937'
-                  }}
-                  className="print-only"
-                >
-                  {formatDateForDisplay(invoiceDate)}
-                </span>
-              </div>
-            </div>
-            <div style={{ 
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'flex-end',
-              gap: '12px'
-            }}>
-              <span
-                role="textbox"
-                aria-label="Due date label"
-                tabIndex={0}
-                contentEditable
-                suppressContentEditableWarning
-                style={{
-                  fontSize: '15px',
-                  fontWeight: '600',
-                  color: '#374151',
-                  outline: 'none',
-                  cursor: 'text',
-                  padding: '6px 10px',
-                  borderRadius: '8px',
-                  transition: 'all 0.2s ease',
-                  backgroundColor: '#f8fafc',
-                  border: '2px solid #e2e8f0',
-                  minWidth: '110px',
-                  width: '110px',
-                  display: 'inline-block',
-                  textAlign: 'right',
-                  whiteSpace: 'nowrap'
-                }}
-                onFocus={(e) => {
-                  e.target.style.backgroundColor = '#ffffff'
-                  e.target.style.borderColor = '#667eea'
-                  e.target.style.boxShadow = '0 0 0 3px rgba(102, 126, 234, 0.1)'
-                }}
-                onBlur={(e) => {
-                  setDueDateLabel(e.target.textContent || 'Due Date:')
-                  e.target.style.backgroundColor = '#f8fafc'
-                  e.target.style.borderColor = '#e2e8f0'
-                  e.target.style.boxShadow = 'none'
-                }}
-                onKeyDown={(e) => {
-                  if (e.key === 'Enter') {
-                    e.preventDefault()
-                    e.currentTarget.blur()
-                  }
-                }}
-              >
-                {dueDateLabel}
-              </span> 
-              <div style={{ position: 'relative', display: 'inline-block' }}>
-                <input
-                  type="date"
-                  value={dueDate}
-                  onChange={(e) => setDueDate(e.target.value)}
-                  style={{
-                    backgroundColor: '#f8fafc',
-                    padding: '10px 16px',
-                    borderRadius: '8px',
-                    border: '2px solid #e2e8f0',
-                    minWidth: '130px',
-                  width: '130px',
-                    fontSize: '14px',
-                    fontWeight: '500',
-                    outline: 'none',
-                    textAlign: 'right',
-                    cursor: 'pointer',
-                    transition: 'all 0.2s ease',
-                    color: '#1f2937'
-                  }}
-                  onFocus={(e) => {
-                    e.target.style.borderColor = '#667eea'
-                    e.target.style.backgroundColor = '#ffffff'
-                    e.target.style.boxShadow = '0 0 0 3px rgba(102, 126, 234, 0.1)'
-                  }}
-                  onBlur={(e) => {
-                    e.target.style.borderColor = '#e2e8f0'
-                    e.target.style.backgroundColor = '#f8fafc'
-                    e.target.style.boxShadow = 'none'
-                  }}
-                />
-                <span 
-                  style={{
-                    position: 'absolute',
-                    right: '16px',
-                    top: '10px',
-                    pointerEvents: 'none',
-                    fontSize: '14px',
-                    color: '#1f2937'
-                  }}
-                  className="print-only"
-                >
-                  {formatDateForDisplay(dueDate)}
-                </span>
-              </div>
-            </div>
-          </div>
-        </div>
-
-        {/* From/To Details */}
-        <div className="details-grid" style={{
-          display: 'grid',
-          gridTemplateColumns: '1fr 1fr',
-          gap: '32px',
-          marginBottom: '40px'
-        }}>
-          <div className="contractor-details" style={{
-            backgroundColor: '#f8fafc',
-            borderRadius: '12px',
-            padding: '24px',
-            border: '1px solid #e2e8f0'
-          }}>
-            <h3 style={{ 
-              color: '#1f2937', 
-              marginBottom: '20px', 
-              fontSize: '18px',
-              fontWeight: '700',
-              display: 'flex',
-              alignItems: 'center',
-              gap: '8px'
-            }}>
-              <span style={{
-                width: '6px',
-                height: '6px',
-                borderRadius: '50%',
-                background: 'linear-gradient(135deg, #667eea 0%, #764ba2 100%)'
-              }}></span>
-              <span
-                role="textbox"
-                aria-label="From section title"
-                tabIndex={0}
-                contentEditable
-                suppressContentEditableWarning
-                style={{
-                  outline: 'none',
-                  cursor: 'text',
-                  padding: '6px 10px',
-                  borderRadius: '8px',
-                  transition: 'all 0.2s ease',
-                  backgroundColor: '#f8fafc',
-                  border: '2px solid #e2e8f0',
-                  minWidth: '60px'
-                }}
-                onFocus={(e) => {
-                  e.target.style.backgroundColor = '#ffffff'
-                  e.target.style.borderColor = '#667eea'
-                  e.target.style.boxShadow = '0 0 0 3px rgba(102, 126, 234, 0.1)'
-                }}
-                onBlur={(e) => {
-                  setFromTitle(e.target.textContent || 'From:')
-                  e.target.style.backgroundColor = '#f8fafc'
-                  e.target.style.borderColor = '#e2e8f0'
-                  e.target.style.boxShadow = 'none'
-                }}
-                onKeyDown={(e) => {
-                  if (e.key === 'Enter') {
-                    e.preventDefault()
-                    e.currentTarget.blur()
-                  }
-                }}
-              >
-                {fromTitle}
-              </span>
-            </h3>
-            <div style={{ display: 'flex', flexDirection: 'column', gap: '12px' }}>
-              {fromFields.map((field) => (
-                <div key={field.id} style={{ display: 'flex', alignItems: 'center', gap: '12px' }}>
-                  <input
-                    type="text"
-                    value={field.value}
-                    placeholder={field.placeholder}
-                    onChange={(e) => updateFromField(field.id, e.target.value)}
-                    autoComplete="off"
-                    style={{
-                      flex: 1,
-                      backgroundColor: '#ffffff',
-                      border: '2px solid #e2e8f0',
-                      borderRadius: '8px',
-                      padding: '12px 16px',
-                      fontSize: '14px',
-                      outline: 'none',
-                      transition: 'all 0.2s ease',
-                      fontWeight: '500',
-                      color: '#1f2937'
-                    }}
-                    onFocus={(e) => {
-                      e.target.style.borderColor = '#667eea'
-                      e.target.style.boxShadow = '0 0 0 3px rgba(102, 126, 234, 0.1)'
-                    }}
-                    onBlur={(e) => {
-                      e.target.style.borderColor = '#e2e8f0'
-                      e.target.style.boxShadow = 'none'
-                    }}
-                  />
-                  {fromFields.length > 1 && (
-                    <button 
-                      className="no-print remove-btn"
-                      onClick={() => removeFromField(field.id)}
-                      style={{
-                        width: '32px',
-                        height: '32px',
-                        backgroundColor: '#fef2f2',
-                        color: '#dc2626',
-                        border: '2px solid #fecaca',
-                        borderRadius: '8px',
-                        cursor: 'pointer',
-                        fontSize: '16px',
-                        display: 'flex',
-                        alignItems: 'center',
-                        justifyContent: 'center',
-                        lineHeight: 1,
-                        transition: 'all 0.2s ease',
-                        fontWeight: '600'
-                      }}
-                      onMouseEnter={(e) => {
-                        e.currentTarget.style.backgroundColor = '#dc2626'
-                        e.currentTarget.style.color = 'white'
-                        e.currentTarget.style.borderColor = '#dc2626'
-                        e.currentTarget.style.transform = 'scale(1.05)'
-                      }}
-                      onMouseLeave={(e) => {
-                        e.currentTarget.style.backgroundColor = '#fef2f2'
-                        e.currentTarget.style.color = '#dc2626'
-                        e.currentTarget.style.borderColor = '#fecaca'
-                        e.currentTarget.style.transform = 'scale(1)'
-                      }}
-                      title="Remove field"
-                    >×</button>
-                  )}
-                </div>
-              ))}
-              <button 
-                className="no-print add-btn"
-                onClick={addFromField}
-                style={{
-                  width: '100%',
-                  height: '40px',
-                  backgroundColor: '#f0f9ff',
-                  color: '#0284c7',
-                  border: '2px dashed #0284c7',
-                  borderRadius: '8px',
-                  cursor: 'pointer',
-                  fontSize: '14px',
-                  fontWeight: '600',
-                  display: 'flex',
-                  alignItems: 'center',
-                  justifyContent: 'center',
-                  gap: '6px',
-                  lineHeight: 1,
-                  marginTop: '8px',
-                  transition: 'all 0.2s ease'
-                }}
-                onMouseEnter={(e) => {
-                  e.currentTarget.style.backgroundColor = '#0284c7'
-                  e.currentTarget.style.color = 'white'
-                  e.currentTarget.style.borderColor = '#0284c7'
-                  e.currentTarget.style.borderStyle = 'solid'
-                  e.currentTarget.style.transform = 'translateY(-1px)'
-                }}
-                onMouseLeave={(e) => {
-                  e.currentTarget.style.backgroundColor = '#f0f9ff'
-                  e.currentTarget.style.color = '#0284c7'
-                  e.currentTarget.style.borderColor = '#0284c7'
-                  e.currentTarget.style.borderStyle = 'dashed'
-                  e.currentTarget.style.transform = 'translateY(0)'
-                }}
-                title="Add field"
-              >
-                + Add Field
-              </button>
-            </div>
-          </div>
-
-          <div className="company-details" style={{
-            backgroundColor: '#f8fafc',
-            borderRadius: '12px',
-            padding: '24px',
-            border: '1px solid #e2e8f0'
-          }}>
-            <h3 style={{ 
-              color: '#1f2937', 
-              marginBottom: '20px', 
-              fontSize: '18px',
-              fontWeight: '700',
-              display: 'flex',
-              alignItems: 'center',
-              gap: '8px'
-            }}>
-              <span style={{
-                width: '6px',
-                height: '6px',
-                borderRadius: '50%',
-                background: 'linear-gradient(135deg, #667eea 0%, #764ba2 100%)'
-              }}></span>
-              <span
-                role="textbox"
-                aria-label="To section title"
-                tabIndex={0}
-                contentEditable
-                suppressContentEditableWarning
-                style={{
-                  outline: 'none',
-                  cursor: 'text',
-                  padding: '6px 10px',
-                  borderRadius: '8px',
-                  transition: 'all 0.2s ease',
-                  backgroundColor: '#f8fafc',
-                  border: '2px solid #e2e8f0',
-                  minWidth: '60px'
-                }}
-                onFocus={(e) => {
-                  e.target.style.backgroundColor = '#ffffff'
-                  e.target.style.borderColor = '#667eea'
-                  e.target.style.boxShadow = '0 0 0 3px rgba(102, 126, 234, 0.1)'
-                }}
-                onBlur={(e) => {
-                  setToTitle(e.target.textContent || 'To:')
-                  e.target.style.backgroundColor = '#f8fafc'
-                  e.target.style.borderColor = '#e2e8f0'
-                  e.target.style.boxShadow = 'none'
-                }}
-                onKeyDown={(e) => {
-                  if (e.key === 'Enter') {
-                    e.preventDefault()
-                    e.currentTarget.blur()
-                  }
-                }}
-              >
-                {toTitle}
-              </span>
-            </h3>
-            <div style={{ display: 'flex', flexDirection: 'column', gap: '12px' }}>
-              {toFields.map((field) => (
-                <div key={field.id} style={{ display: 'flex', alignItems: 'center', gap: '12px' }}>
-                  <input
-                    type="text"
-                    value={field.value}
-                    placeholder={field.placeholder}
-                    onChange={(e) => updateToField(field.id, e.target.value)}
-                    autoComplete="off"
-                    style={{
-                      flex: 1,
-                      backgroundColor: '#ffffff',
-                      border: '2px solid #e2e8f0',
-                      borderRadius: '8px',
-                      padding: '12px 16px',
-                      fontSize: '14px',
-                      outline: 'none',
-                      transition: 'all 0.2s ease',
-                      fontWeight: '500',
-                      color: '#1f2937'
-                    }}
-                    onFocus={(e) => {
-                      e.target.style.borderColor = '#667eea'
-                      e.target.style.boxShadow = '0 0 0 3px rgba(102, 126, 234, 0.1)'
-                    }}
-                    onBlur={(e) => {
-                      e.target.style.borderColor = '#e2e8f0'
-                      e.target.style.boxShadow = 'none'
-                    }}
-                  />
-                  {toFields.length > 1 && (
-                    <button 
-                      className="no-print remove-btn"
-                      onClick={() => removeToField(field.id)}
-                      style={{
-                        width: '32px',
-                        height: '32px',
-                        backgroundColor: '#fef2f2',
-                        color: '#dc2626',
-                        border: '2px solid #fecaca',
-                        borderRadius: '8px',
-                        cursor: 'pointer',
-                        fontSize: '16px',
-                        display: 'flex',
-                        alignItems: 'center',
-                        justifyContent: 'center',
-                        lineHeight: 1,
-                        transition: 'all 0.2s ease',
-                        fontWeight: '600'
-                      }}
-                      onMouseEnter={(e) => {
-                        e.currentTarget.style.backgroundColor = '#dc2626'
-                        e.currentTarget.style.color = 'white'
-                        e.currentTarget.style.borderColor = '#dc2626'
-                        e.currentTarget.style.transform = 'scale(1.05)'
-                      }}
-                      onMouseLeave={(e) => {
-                        e.currentTarget.style.backgroundColor = '#fef2f2'
-                        e.currentTarget.style.color = '#dc2626'
-                        e.currentTarget.style.borderColor = '#fecaca'
-                        e.currentTarget.style.transform = 'scale(1)'
-                      }}
-                      title="Remove field"
-                    >×</button>
-                  )}
-                </div>
-              ))}
-              <button 
-                className="no-print add-btn"
-                onClick={addToField}
-                style={{
-                  width: '100%',
-                  height: '40px',
-                  backgroundColor: '#f0f9ff',
-                  color: '#0284c7',
-                  border: '2px dashed #0284c7',
-                  borderRadius: '8px',
-                  cursor: 'pointer',
-                  fontSize: '14px',
-                  fontWeight: '600',
-                  display: 'flex',
-                  alignItems: 'center',
-                  justifyContent: 'center',
-                  gap: '6px',
-                  lineHeight: 1,
-                  marginTop: '8px',
-                  transition: 'all 0.2s ease'
-                }}
-                onMouseEnter={(e) => {
-                  e.currentTarget.style.backgroundColor = '#0284c7'
-                  e.currentTarget.style.color = 'white'
-                  e.currentTarget.style.borderColor = '#0284c7'
-                  e.currentTarget.style.borderStyle = 'solid'
-                  e.currentTarget.style.transform = 'translateY(-1px)'
-                }}
-                onMouseLeave={(e) => {
-                  e.currentTarget.style.backgroundColor = '#f0f9ff'
-                  e.currentTarget.style.color = '#0284c7'
-                  e.currentTarget.style.borderColor = '#0284c7'
-                  e.currentTarget.style.borderStyle = 'dashed'
-                  e.currentTarget.style.transform = 'translateY(0)'
-                }}
-                title="Add field"
-              >
-                + Add Field
-              </button>
-            </div>
-          </div>
-        </div>
-
-        {/* Services Table */}
-        <div style={{ 
-          marginBottom: '40px',
-          borderRadius: '12px',
-          overflow: 'hidden',
-          border: '1px solid #e2e8f0',
-          marginRight: '0px',
-          position: 'relative'
-        }}>
-        <table className="services-table" style={{
-          width: '100%',
-          borderCollapse: 'collapse'
-        }}>
-          <thead>
-            <tr style={{
-              background: 'linear-gradient(135deg, #f8fafc 0%, #f1f5f9 100%)',
-              position: 'relative'
-            }}>
-              {columns.map((column, index) => (
-                <React.Fragment key={column.id}>
-                  <th style={{
-                    padding: '20px 16px',
-                    textAlign: 'center',
-                    fontWeight: '700',
-                    color: '#1f2937',
-                    width: column.width,
-                    position: 'relative',
-                    fontSize: '15px'
-                  }}>
-                    {/* Add button between columns (appears on hover) - allows adding before Amount */}
-                    {index > 0 && (
-                      <div 
-                        className="no-print"
-                        style={{
-                          position: 'absolute',
-                          left: '-20px',
-                          top: '0',
-                          bottom: '0',
-                          width: '40px',
-                          display: 'flex',
-                          alignItems: 'center',
-                          justifyContent: 'center',
-                          zIndex: 10
-                        }}
-                        onMouseEnter={(e) => {
-                          const btn = e.currentTarget.querySelector('button')
-                          if (btn) btn.style.opacity = '1'
-                        }}
-                        onMouseLeave={(e) => {
-                          const btn = e.currentTarget.querySelector('button')
-                          if (btn) btn.style.opacity = '0'
-                        }}
-                      >
-                        <button 
-                          onClick={() => addColumn(columns[index - 1].id)}
-                          style={{
-                            width: '24px',
-                            height: '24px',
-                            backgroundColor: '#0284c7',
-                            color: 'white',
-                            border: 'none',
-                            borderRadius: '50%',
-                            cursor: 'pointer',
-                            fontSize: '16px',
-                            display: 'flex',
-                            alignItems: 'center',
-                            justifyContent: 'center',
-                            lineHeight: 1,
-                            boxShadow: '0 2px 4px rgba(0,0,0,0.1)',
-                            fontWeight: '600',
-                            opacity: 0,
-                            transition: 'opacity 0.2s ease'
-                          }}
-                          title="Add column here"
-                        >+</button>
-                      </div>
-                    )}
-                    
-                    <span
-                      role="textbox"
-                      aria-label={`Column header: ${column.name}`}
-                      tabIndex={0}
-                      contentEditable 
-                      suppressContentEditableWarning
-                      style={{
-                        backgroundColor: '#ffffff',
-                        padding: '8px 12px',
-                        borderRadius: '6px',
-                        border: '2px solid #e2e8f0',
-                        minWidth: '80px',
-                        display: 'inline-block',
-                        transition: 'all 0.2s ease',
-                        cursor: 'text'
-                      }}
-                      onFocus={(e) => {
-                        e.target.style.borderColor = '#667eea'
-                        e.target.style.boxShadow = '0 0 0 3px rgba(102, 126, 234, 0.1)'
-                      }}
-                      onBlur={(e) => {
-                        updateColumnName(column.id, e.target.textContent || column.name)
-                        e.target.style.borderColor = '#e2e8f0'
-                        e.target.style.boxShadow = 'none'
-                      }}
-                      onKeyDown={(e) => {
-                        if (e.key === 'Enter') {
-                          e.preventDefault()
-                          e.currentTarget.blur()
-                        }
-                      }}
-                    >
-                      {column.name}
-                    </span>
-                    
-                    {/* Delete button - now centered horizontally */}
-                    {!column.isDescription && columns.length > 2 && (
-                      <button 
-                        className="no-print remove-btn"
-                        onClick={() => removeColumn(column.id)}
-                        style={{
-                          position: 'absolute',
-                          top: '4px',
-                          left: '50%',
-                          transform: 'translateX(-50%)',
-                          width: '24px',
-                          height: '24px',
-                          backgroundColor: '#fef2f2',
-                          color: '#dc2626',
-                          border: '2px solid #fecaca',
-                          borderRadius: '6px',
-                          cursor: 'pointer',
-                          fontSize: '14px',
-                          display: 'flex',
-                          alignItems: 'center',
-                          justifyContent: 'center',
-                          lineHeight: 1,
-                          transition: 'all 0.2s ease',
-                          fontWeight: '600',
-                          marginLeft: '0',
-                          marginRight: '0',
-                          boxSizing: 'border-box'
-                        }}
-                        onMouseEnter={(e) => {
-                          e.currentTarget.style.backgroundColor = '#dc2626'
-                          e.currentTarget.style.color = 'white'
-                          e.currentTarget.style.borderColor = '#dc2626'
-                          e.currentTarget.style.transform = 'translateX(-50%) scale(1.05)'
-                        }}
-                        onMouseLeave={(e) => {
-                          e.currentTarget.style.backgroundColor = '#fef2f2'
-                          e.currentTarget.style.color = '#dc2626'
-                          e.currentTarget.style.borderColor = '#fecaca'
-                          e.currentTarget.style.transform = 'translateX(-50%) scale(1)'
-                        }}
-                        title="Remove column"
-                      >×</button>
-                    )}
-                  </th>
-                  
-                  {/* Add button after the last column - only if Amount is not the last column */}
-                  {index === columns.length - 1 && !column.isAmount && (
-                    <div 
-                      className="no-print"
-                      style={{
-                        position: 'absolute',
-                        right: '-20px',
-                        top: '0',
-                        bottom: '0',
-                        width: '40px',
-                        display: 'flex',
-                        alignItems: 'center',
-                        justifyContent: 'center',
-                        zIndex: 10
-                      }}
-                      onMouseEnter={(e) => {
-                        const btn = e.currentTarget.querySelector('button')
-                        if (btn) btn.style.opacity = '1'
-                      }}
-                      onMouseLeave={(e) => {
-                        const btn = e.currentTarget.querySelector('button')
-                        if (btn) btn.style.opacity = '0'
-                      }}
-                    >
-                      <button 
-                        onClick={() => addColumn(column.id)}
-                        style={{
-                          width: '24px',
-                          height: '24px',
-                          backgroundColor: '#0284c7',
-                          color: 'white',
-                          border: 'none',
-                          borderRadius: '50%',
-                          cursor: 'pointer',
-                          fontSize: '16px',
-                          display: 'flex',
-                          alignItems: 'center',
-                          justifyContent: 'center',
-                          lineHeight: 1,
-                          boxShadow: '0 2px 4px rgba(0,0,0,0.1)',
-                          fontWeight: '600',
-                          opacity: 0,
-                          transition: 'opacity 0.2s ease'
-                        }}
-                        title="Add column here"
-                      >+</button>
-                    </div>
-                  )}
-                </React.Fragment>
-              ))}
-            </tr>
-          </thead>
-          <tbody>
-            {rows.map((row, index) => (
-              <tr key={row.id} style={{ 
-                height: '80px',
-                backgroundColor: index % 2 === 0 ? '#ffffff' : '#fafbfc',
-                position: 'relative'
-              }}>
-                {columns.map((column) => (
-                  <td key={`${row.id}-${column.id}`} style={{
-                    width: column.width,
-                    textAlign: column.align as 'left' | 'center' | 'right',
-                    verticalAlign: 'top',
-                    padding: '16px',
-                    borderBottom: '1px solid #f1f3f5'
-                  }}>
-                    {renderCellContent(row, column)}
-                  </td>
-                ))}
-                {/* Remove row button - positioned absolutely to avoid table structure issues */}
-                {rows.length > 1 && (
-                  <div 
-                    className="no-print"
-                    style={{
-                      position: 'absolute',
-                      right: '-20px',
-                      top: '50%',
-                      transform: 'translateY(-50%)',
-                      zIndex: 10
-                    }}
-                  >
-                    <button 
-                      className="remove-btn"
-                      onClick={() => removeRow(row.id)}
-                      style={{
-                        width: '24px',
-                        height: '24px',
-                        backgroundColor: '#fef2f2',
-                        color: '#dc2626',
-                        border: '2px solid #fecaca',
-                        borderRadius: '6px',
-                        cursor: 'pointer',
-                        fontSize: '14px',
-                        display: 'flex',
-                        alignItems: 'center',
-                        justifyContent: 'center',
-                        lineHeight: 1,
-                        transition: 'all 0.2s ease',
-                        fontWeight: '600'
-                      }}
-                      onMouseEnter={(e) => {
-                        e.currentTarget.style.backgroundColor = '#dc2626'
-                        e.currentTarget.style.color = 'white'
-                        e.currentTarget.style.borderColor = '#dc2626'
-                        e.currentTarget.style.transform = 'scale(1.05)'
-                      }}
-                      onMouseLeave={(e) => {
-                        e.currentTarget.style.backgroundColor = '#fef2f2'
-                        e.currentTarget.style.color = '#dc2626'
-                        e.currentTarget.style.borderColor = '#fecaca'
-                        e.currentTarget.style.transform = 'scale(1)'
-                      }}
-                      title="Remove row"
-                    >×</button>
-                  </div>
-                )}
-              </tr>
-            ))}
-          </tbody>
-        </table>
-
-        </div>
-
-        {/* Add Row Button */}
-        <div className="no-print" style={{ marginBottom: '16px', marginTop: '-8px' }}>
-          <button 
-            onClick={addRow}
-            style={{
-              width: '120px',
-              height: '40px',
-              backgroundColor: '#f0f9ff',
-              color: '#0284c7',
-              border: '2px dashed #0284c7',
-              borderRadius: '8px',
-              cursor: 'pointer',
-              fontSize: '14px',
-              fontWeight: '600',
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'center',
-              gap: '6px',
-              lineHeight: 1,
-              transition: 'all 0.2s ease'
-            }}
-            onMouseEnter={(e) => {
-              e.currentTarget.style.backgroundColor = '#0284c7'
-              e.currentTarget.style.color = 'white'
-              e.currentTarget.style.borderColor = '#0284c7'
-              e.currentTarget.style.borderStyle = 'solid'
-              e.currentTarget.style.transform = 'translateY(-1px)'
-            }}
-            onMouseLeave={(e) => {
-              e.currentTarget.style.backgroundColor = '#f0f9ff'
-              e.currentTarget.style.color = '#0284c7'
-              e.currentTarget.style.borderColor = '#0284c7'
-              e.currentTarget.style.borderStyle = 'dashed'
-              e.currentTarget.style.transform = 'translateY(0)'
-            }}
-            title="Add new row"
-          >
-            + Add Row
-          </button>
-        </div>
-
-        {/* Currency Selector */}
-        <div className="currency-section no-print" style={{
-          marginTop: '20px',
-          marginBottom: '20px',
-          padding: '16px',
-          backgroundColor: '#f8fafc',
-          borderRadius: '12px',
-          border: '1px solid #e2e8f0',
-          display: 'flex',
-          alignItems: 'center',
-          gap: '12px'
-        }}>
-          <span style={{
-            fontSize: '15px',
-            fontWeight: '600',
-            color: '#374151'
-          }}>Currency:</span>
-          <CurrencySelector
-            value={currencyCode}
-            symbol={currencySymbol}
-            onChange={(code, symbol) => {
-              setCurrencyCode(code)
-              setCurrencySymbol(symbol)
-            }}
+          
+          {/* Invoice Header */}
+          <InvoiceHeader
+            invoiceTitle={invoiceTitle}
+            onInvoiceTitleChange={setInvoiceTitle}
+            invoiceNumberLabel={invoiceNumberLabel}
+            onInvoiceNumberLabelChange={setInvoiceNumberLabel}
+            dateLabel={dateLabel}
+            onDateLabelChange={setDateLabel}
+            dueDateLabel={dueDateLabel}
+            onDueDateLabelChange={setDueDateLabel}
+            invoiceDate={invoiceDate}
+            onInvoiceDateChange={setInvoiceDate}
+            dueDate={dueDate}
+            onDueDateChange={setDueDate}
           />
-        </div>
 
-        {/* Totals Section */}
-        <div className="total-section" style={{
-          marginTop: '20px',
-          paddingTop: '20px',
-          borderTop: '2px solid #e9ecef',
-          marginBottom: '20px'
-        }}>
-          <div className="total-row" style={{
-            display: 'flex',
-            justifyContent: 'space-between',
-            marginBottom: '10px'
-          }}>
-            <span>Subtotal:</span>
-            <span style={{
-              textAlign: 'right',
-              display: 'inline-block',
-              minWidth: '100px',
-              backgroundColor: '#f8f9fa',
-              padding: '2px 4px',
-              borderRadius: '3px',
-              border: '1px solid #e9ecef'
-            }}>{currencySymbol}{calculateSubtotal()}</span>
-          </div>
+          {/* From/To Details */}
+          <ContactDetails
+            fromTitle={fromTitle}
+            onFromTitleChange={setFromTitle}
+            fromFields={fromFields}
+            onFromFieldsChange={setFromFields}
+            toTitle={toTitle}
+            onToTitleChange={setToTitle}
+            toFields={toFields}
+            onToFieldsChange={setToFields}
+          />
 
-          {/* Discount Row - Only show when enabled */}
-          {discountEnabled && (
-            <div className="total-row" style={{
-              display: 'flex',
-              justifyContent: 'space-between',
-              alignItems: 'center',
-              marginBottom: '10px'
-            }}>
-              <div style={{ display: 'flex', alignItems: 'center', gap: '8px' }}>
-                <span>Discount ({discountPercentage}%):</span>
-              </div>
-              <span style={{
-                textAlign: 'right',
-                display: 'inline-block',
-                minWidth: '100px',
-                backgroundColor: '#f8f9fa',
-                padding: '2px 4px',
-                borderRadius: '3px',
-                border: '1px solid #e9ecef'
-              }}>({currencySymbol}{calculateDiscountAmount()})</span>
-            </div>
-          )}
+          {/* Services Table */}
+          <ServicesTable
+            columns={columns}
+            rows={rows}
+            onColumnsChange={setColumns}
+            onRowsChange={setRows}
+            onCellUpdate={handleCellUpdate}
+          />
 
-          {/* Tax Row - Only show when enabled */}
-          {taxEnabled && (
-            <div className="total-row" style={{
-              display: 'flex',
-              justifyContent: 'space-between',
-              alignItems: 'center',
-              marginBottom: '10px'
-            }}>
-              <div style={{ display: 'flex', alignItems: 'center', gap: '8px' }}>
-                <span>Tax ({taxPercentage}%):</span>
-              </div>
-              <span style={{
-                textAlign: 'right',
-                display: 'inline-block',
-                minWidth: '100px',
-                backgroundColor: '#f8f9fa',
-                padding: '2px 4px',
-                borderRadius: '3px',
-                border: '1px solid #e9ecef'
-              }}>{currencySymbol}{calculateTaxAmount()}</span>
-            </div>
-          )}
-
-          {/* Control Panel for Tax/Discount (only visible when editing) */}
-          <div className="no-print" style={{
-            display: 'flex',
-            gap: '20px',
-            marginBottom: '10px',
+          {/* Currency Selector */}
+          <div className="currency-section no-print" style={{
+            marginTop: '20px',
+            marginBottom: '20px',
             padding: '16px',
             backgroundColor: '#f8fafc',
-            borderRadius: '8px',
-            fontSize: '14px',
-            border: '1px solid #e2e8f0'
-          }}>
-            <div style={{ display: 'flex', alignItems: 'center', gap: '8px' }}>
-              <input
-                type="checkbox"
-                checked={discountEnabled}
-                onChange={(e) => setDiscountEnabled(e.target.checked)}
-                style={{ cursor: 'pointer', width: '16px', height: '16px' }}
-              />
-              <span style={{ color: discountEnabled ? '#1f2937' : '#9ca3af', fontWeight: '500' }}>Add Discount:</span>
-              <input
-                type="text"
-                value={discountPercentage}
-                onChange={(e) => setDiscountPercentage(e.target.value)}
-                disabled={!discountEnabled}
-                style={{
-                  width: '60px',
-                  backgroundColor: discountEnabled ? '#ffffff' : '#f3f4f6',
-                  border: discountEnabled ? '2px solid #e2e8f0' : '2px solid #e5e7eb',
-                  borderRadius: '6px',
-                  padding: '6px 8px',
-                  fontSize: '14px',
-                  textAlign: 'center',
-                  color: discountEnabled ? '#1f2937' : '#9ca3af',
-                  fontWeight: '500',
-                  transition: 'all 0.2s ease',
-                  outline: 'none'
-                }}
-                onFocus={(e) => {
-                  if (discountEnabled) {
-                    e.target.style.borderColor = '#667eea'
-                    e.target.style.boxShadow = '0 0 0 3px rgba(102, 126, 234, 0.1)'
-                  }
-                }}
-                onBlur={(e) => {
-                  if (discountEnabled) {
-                    e.target.style.borderColor = '#e2e8f0'
-                    e.target.style.boxShadow = 'none'
-                  }
-                }}
-              />
-              <span style={{ color: discountEnabled ? '#1f2937' : '#9ca3af', fontWeight: '500' }}>%</span>
-            </div>
-            
-            <div style={{ display: 'flex', alignItems: 'center', gap: '8px' }}>
-              <input
-                type="checkbox"
-                checked={taxEnabled}
-                onChange={(e) => setTaxEnabled(e.target.checked)}
-                style={{ cursor: 'pointer', width: '16px', height: '16px' }}
-              />
-              <span style={{ color: taxEnabled ? '#1f2937' : '#9ca3af', fontWeight: '500' }}>Add Tax:</span>
-              <input
-                type="text"
-                value={taxPercentage}
-                onChange={(e) => setTaxPercentage(e.target.value)}
-                disabled={!taxEnabled}
-                style={{
-                  width: '60px',
-                  backgroundColor: taxEnabled ? '#ffffff' : '#f3f4f6',
-                  border: taxEnabled ? '2px solid #e2e8f0' : '2px solid #e5e7eb',
-                  borderRadius: '6px',
-                  padding: '6px 8px',
-                  fontSize: '14px',
-                  textAlign: 'center',
-                  color: taxEnabled ? '#1f2937' : '#9ca3af',
-                  fontWeight: '500',
-                  transition: 'all 0.2s ease',
-                  outline: 'none'
-                }}
-                onFocus={(e) => {
-                  if (taxEnabled) {
-                    e.target.style.borderColor = '#667eea'
-                    e.target.style.boxShadow = '0 0 0 3px rgba(102, 126, 234, 0.1)'
-                  }
-                }}
-                onBlur={(e) => {
-                  if (taxEnabled) {
-                    e.target.style.borderColor = '#e2e8f0'
-                    e.target.style.boxShadow = 'none'
-                  }
-                }}
-              />
-              <span style={{ color: taxEnabled ? '#1f2937' : '#9ca3af', fontWeight: '500' }}>%</span>
-            </div>
-          </div>
-
-          <div className="total-row total-final" style={{
-            display: 'flex',
-            justifyContent: 'space-between',
-            fontWeight: 'bold',
-            fontSize: '18px',
-            color: '#2c3e50',
-            borderTop: '1px solid #ccc',
-            paddingTop: '10px'
-          }}>
-            <span
-              role="textbox"
-              aria-label="Total amount label"
-              tabIndex={0}
-              contentEditable
-              suppressContentEditableWarning
-              style={{
-                outline: 'none',
-                cursor: 'text',
-                padding: '6px 10px',
-                borderRadius: '8px',
-                transition: 'all 0.2s ease',
-                backgroundColor: '#f8fafc',
-                border: '2px solid #e2e8f0',
-                minWidth: '140px',
-                display: 'inline-block'
-              }}
-              onFocus={(e) => {
-                e.target.style.backgroundColor = '#ffffff'
-                e.target.style.borderColor = '#667eea'
-                e.target.style.boxShadow = '0 0 0 3px rgba(102, 126, 234, 0.1)'
-              }}
-              onBlur={(e) => {
-                setTotalLabel(e.target.textContent || 'Total Amount Due:')
-                e.target.style.backgroundColor = '#f8fafc'
-                e.target.style.borderColor = '#e2e8f0'
-                e.target.style.boxShadow = 'none'
-              }}
-              onKeyDown={(e) => {
-                if (e.key === 'Enter') {
-                  e.preventDefault()
-                  e.currentTarget.blur()
-                }
-              }}
-            >
-              {totalLabel}
-            </span>
-            <span style={{
-              textAlign: 'right',
-              display: 'inline-block',
-              minWidth: '100px',
-              backgroundColor: '#f8f9fa',
-              padding: '2px 4px',
-              borderRadius: '3px',
-              border: '1px solid #e9ecef',
-              fontWeight: 'bold'
-            }}>{currencySymbol}{calculateTotal()}</span>
-          </div>
-        </div>
-
-        {/* Bank Details */}
-        <div className="bank-details" style={{
-          backgroundColor: '#f8fafc',
-          border: '1px solid #e2e8f0',
-          padding: '24px',
-          borderRadius: '12px',
-          marginTop: '0px'
-        }}>
-          <h4 style={{ 
-            color: '#1f2937', 
-            marginBottom: '20px',
-            fontSize: '18px',
-            fontWeight: '700',
+            borderRadius: '12px',
+            border: '1px solid #e2e8f0',
             display: 'flex',
             alignItems: 'center',
-            gap: '8px'
+            gap: '12px'
           }}>
             <span style={{
-              width: '6px',
-              height: '6px',
-              borderRadius: '50%',
-              background: 'linear-gradient(135deg, #667eea 0%, #764ba2 100%)'
-            }}></span>
-            <span
-              role="textbox"
-              aria-label="Payment section title"
-              tabIndex={0}
-              contentEditable
-              suppressContentEditableWarning
-              style={{
-                outline: 'none',
-                cursor: 'text',
-                padding: '6px 10px',
-                borderRadius: '8px',
-                transition: 'all 0.2s ease',
-                backgroundColor: '#f8fafc',
-                border: '2px solid #e2e8f0',
-                minWidth: '140px'
-              }}
-              onFocus={(e) => {
-                e.target.style.backgroundColor = '#ffffff'
-                e.target.style.borderColor = '#667eea'
-                e.target.style.boxShadow = '0 0 0 3px rgba(102, 126, 234, 0.1)'
-              }}
-              onBlur={(e) => {
-                setPaymentTitle(e.target.textContent || 'Payment Details:')
-                e.target.style.backgroundColor = '#f8fafc'
-                e.target.style.borderColor = '#e2e8f0'
-                e.target.style.boxShadow = 'none'
-              }}
-              onKeyDown={(e) => {
-                if (e.key === 'Enter') {
-                  e.preventDefault()
-                  e.currentTarget.blur()
-                }
-              }}
-            >
-              {paymentTitle}
-            </span>
-          </h4>
-          <div style={{ 
-            display: 'grid',
-            gridTemplateColumns: '1fr 1fr',
-            gap: '30px'
-          }}>
-            <div style={{ display: 'flex', flexDirection: 'column', gap: '12px' }}>
-              {paymentFields.map((field) => (
-                <div key={field.id} style={{ display: 'flex', alignItems: 'center', gap: '12px' }}>
-                  <input
-                    type="text"
-                    value={field.value}
-                    placeholder={field.placeholder}
-                    onChange={(e) => updatePaymentField(field.id, e.target.value)}
-                    autoComplete="off"
-                    style={{
-                      flex: 1,
-                      backgroundColor: '#ffffff',
-                      border: '2px solid #e2e8f0',
-                      borderRadius: '8px',
-                      padding: '12px 16px',
-                      fontSize: '14px',
-                      outline: 'none',
-                      transition: 'all 0.2s ease',
-                      fontWeight: '500',
-                      color: '#1f2937'
-                    }}
-                    onFocus={(e) => {
-                      e.target.style.borderColor = '#667eea'
-                      e.target.style.boxShadow = '0 0 0 3px rgba(102, 126, 234, 0.1)'
-                    }}
-                    onBlur={(e) => {
-                      e.target.style.borderColor = '#e2e8f0'
-                      e.target.style.boxShadow = 'none'
-                    }}
-                  />
-                  {paymentFields.length > 1 && (
-                    <button 
-                      className="no-print remove-btn"
-                      onClick={() => removePaymentField(field.id)}
-                      style={{
-                        width: '32px',
-                        height: '32px',
-                        backgroundColor: '#fef2f2',
-                        color: '#dc2626',
-                        border: '2px solid #fecaca',
-                        borderRadius: '8px',
-                        cursor: 'pointer',
-                        fontSize: '16px',
-                        display: 'flex',
-                        alignItems: 'center',
-                        justifyContent: 'center',
-                        lineHeight: 1,
-                        transition: 'all 0.2s ease',
-                        fontWeight: '600'
-                      }}
-                      onMouseEnter={(e) => {
-                        e.currentTarget.style.backgroundColor = '#dc2626'
-                        e.currentTarget.style.color = 'white'
-                        e.currentTarget.style.borderColor = '#dc2626'
-                        e.currentTarget.style.transform = 'scale(1.05)'
-                      }}
-                      onMouseLeave={(e) => {
-                        e.currentTarget.style.backgroundColor = '#fef2f2'
-                        e.currentTarget.style.color = '#dc2626'
-                        e.currentTarget.style.borderColor = '#fecaca'
-                        e.currentTarget.style.transform = 'scale(1)'
-                      }}
-                      title="Remove field"
-                    >×</button>
-                  )}
-                </div>
-              ))}
-              <button 
-                className="no-print add-btn"
-                onClick={addPaymentField}
-                style={{
-                  width: '100%',
-                  height: '40px',
-                  backgroundColor: '#f0f9ff',
-                  color: '#0284c7',
-                  border: '2px dashed #0284c7',
-                  borderRadius: '8px',
-                  cursor: 'pointer',
-                  fontSize: '14px',
-                  fontWeight: '600',
-                  display: 'flex',
-                  alignItems: 'center',
-                  justifyContent: 'center',
-                  gap: '6px',
-                  lineHeight: 1,
-                  marginTop: '8px',
-                  transition: 'all 0.2s ease'
-                }}
-                onMouseEnter={(e) => {
-                  e.currentTarget.style.backgroundColor = '#0284c7'
-                  e.currentTarget.style.color = 'white'
-                  e.currentTarget.style.borderColor = '#0284c7'
-                  e.currentTarget.style.borderStyle = 'solid'
-                  e.currentTarget.style.transform = 'translateY(-1px)'
-                }}
-                onMouseLeave={(e) => {
-                  e.currentTarget.style.backgroundColor = '#f0f9ff'
-                  e.currentTarget.style.color = '#0284c7'
-                  e.currentTarget.style.borderColor = '#0284c7'
-                  e.currentTarget.style.borderStyle = 'dashed'
-                  e.currentTarget.style.transform = 'translateY(0)'
-                }}
-                title="Add field"
-              >
-                + Add Field
-              </button>
-            </div>
-            <div>{/* Empty column for spacing */}</div>
-          </div>
-        </div>
-
-        {/* Action Buttons */}
-        <div className="no-print" style={{
-          display: 'flex',
-          gap: '16px',
-          justifyContent: 'center',
-          marginTop: '40px',
-          paddingTop: '30px',
-          borderTop: '2px solid #f1f3f5'
-        }}>
-          <button 
-            className="print-button"
-            onClick={() => window.print()}
-            style={{
-              background: 'linear-gradient(135deg, #667eea 0%, #764ba2 100%)',
-              color: 'white',
-              padding: '16px 32px',
-              border: 'none',
-              borderRadius: '12px',
-              cursor: 'pointer',
-              fontSize: '16px',
+              fontSize: '15px',
               fontWeight: '600',
-              display: 'flex',
-              alignItems: 'center',
-              gap: '10px',
-              transition: 'all 0.2s ease',
-              boxShadow: '0 6px 20px rgba(102, 126, 234, 0.3)',
-              minWidth: '180px',
-              justifyContent: 'center'
-            }}
-            onMouseEnter={(e) => {
-              e.currentTarget.style.transform = 'translateY(-3px)'
-              e.currentTarget.style.boxShadow = '0 8px 30px rgba(102, 126, 234, 0.4)'
-            }}
-            onMouseLeave={(e) => {
-              e.currentTarget.style.transform = 'translateY(0)'
-              e.currentTarget.style.boxShadow = '0 6px 20px rgba(102, 126, 234, 0.3)'
-            }}
-          >
-            <svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-              <polyline points="6,9 6,2 18,2 18,9"></polyline>
-              <path d="M6,18H4a2,2,0,0,1-2-2V11a2,2,0,0,1,2-2H20a2,2,0,0,1,2,2v5a2,2,0,0,1-2,2H18"></path>
-              <rect x="6" y="14" width="12" height="8"></rect>
-            </svg>
-            Print/Save as PDF
-          </button>
+              color: '#374151'
+            }}>Currency:</span>
+            <CurrencySelector
+              value={currencyCode}
+              symbol={currencySymbol}
+              onChange={(code, symbol) => {
+                setCurrencyCode(code)
+                setCurrencySymbol(symbol)
+              }}
+            />
+          </div>
+
+          {/* Totals Section */}
+          <TotalsSection
+            subtotal={totals.subtotal}
+            tax={totals.tax}
+            discount={totals.discount}
+            total={totals.total}
+            taxEnabled={taxEnabled}
+            taxPercentage={taxPercentage}
+            discountEnabled={discountEnabled}
+            discountPercentage={discountPercentage}
+            currencySymbol={currencySymbol}
+            totalLabel={totalLabel}
+            onTaxToggle={setTaxEnabled}
+            onTaxPercentageChange={setTaxPercentage}
+            onDiscountToggle={setDiscountEnabled}
+            onDiscountPercentageChange={setDiscountPercentage}
+            onTotalLabelChange={setTotalLabel}
+          />
+
+          {/* Bank Details */}
+          <PaymentDetails
+            paymentTitle={paymentTitle}
+            onPaymentTitleChange={setPaymentTitle}
+            paymentFields={paymentFields}
+            onPaymentFieldsChange={setPaymentFields}
+          />
+
+          {/* Action Buttons */}
+          <ActionButtons />
         </div>
-      </div>
       </div>
 
       <style>{`
@@ -2156,77 +409,6 @@ export default function FlexibleInvoice() {
             line-height: 1.2 !important;
           }
           
-          /* Optimize vertical spacing between sections */
-          .invoice-container > div {
-            margin-bottom: 16px !important;
-          }
-          
-          /* Optimize the header section */
-          .invoice-header {
-            padding-bottom: 12px !important;
-            margin-bottom: 16px !important;
-          }
-          
-          /* Optimize From/To sections */
-          .from-to-section {
-            margin-bottom: 16px !important;
-          }
-          
-          .from-to-section h3,
-          .from-to-section h4 {
-            margin-bottom: 6px !important;
-          }
-          
-          .from-to-section div {
-            margin-bottom: 3px !important;
-          }
-          
-          /* Fix header alignment - ensure all labels and inputs align properly */
-          .invoice-details {
-            text-align: right !important;
-          }
-          
-          .invoice-details > div {
-            margin-bottom: 8px !important;
-            display: flex !important;
-            align-items: center !important;
-            justify-content: flex-end !important;
-            gap: 8px !important;
-            height: 24px !important;
-          }
-          
-          /* Ensure consistent label width for proper alignment */
-          .invoice-details span[contenteditable] {
-            min-width: 110px !important;
-            width: 110px !important;
-            text-align: right !important;
-            white-space: nowrap !important;
-            display: inline-flex !important;
-            align-items: center !important;
-            justify-content: flex-end !important;
-            line-height: 1 !important;
-            padding: 0 !important;
-            margin: 0 !important;
-            height: 24px !important;
-          }
-          
-          /* Make all inputs in header have consistent width and proper alignment */
-          .invoice-details input[type="text"],
-          .invoice-details input[type="date"] {
-            width: 130px !important;
-            min-width: 130px !important;
-            max-width: 130px !important;
-            text-align: right !important;
-            padding: 2px 4px !important;
-            margin: 0 !important;
-            line-height: 1 !important;
-            height: 24px !important;
-            display: inline-flex !important;
-            align-items: center !important;
-            font-size: 16px !important;
-            box-sizing: border-box !important;
-          }
-          
           /* Remove grey boxes from totals */
           .total-section span[style*="background"],
           .total-row span[style*="backgroundColor"] {
@@ -2261,101 +443,10 @@ export default function FlexibleInvoice() {
             break-inside: avoid !important;
           }
           
-          /* Remove extra spacing that might cause blank pages */
-          .invoice-container {
-            padding-bottom: 0 !important;
-            margin-bottom: 0 !important;
-          }
-          
           /* Ensure proper page margins and prevent blank pages */
           @page {
             margin: 0.6in !important;
             size: A4 !important;
-          }
-          
-          /* Prevent blank second page - following reference template approach */
-          body {
-            background-color: white !important;
-            padding: 0 !important;
-            margin: 0 !important;
-          }
-          
-          /* Hide the main wrapper that causes extra height */
-          body > main {
-            display: none !important;
-          }
-          
-          .invoice-container {
-            display: block !important;
-            box-shadow: none !important;
-            padding: 25px !important;
-            margin: 0 !important;
-            page-break-after: avoid !important;
-          }
-          
-          /* Ensure no extra height anywhere */
-          html {
-            height: auto !important;
-          }
-          
-          body {
-            height: auto !important;
-          }
-          
-          /* Prevent any element from forcing a page break */
-          * {
-            page-break-after: auto !important;
-          }
-          
-          /* Specifically prevent page break on last element */
-          .invoice-container > *:last-child,
-          .invoice-container > div:last-child,
-          .notes-section,
-          .bank-details {
-            page-break-after: avoid !important;
-            page-break-inside: avoid !important;
-            margin-bottom: 0 !important;
-            padding-bottom: 0 !important;
-          }
-          
-          /* Hide any invisible spacer elements that might cause blank pages */
-          div:empty {
-            display: none !important;
-          }
-          
-          /* Ensure the invoice fits on one page if possible */
-          .invoice-container {
-            max-height: 100vh !important;
-          }
-          
-          /* Ensure the table container spans full width */
-          .services-table-container,
-          div:has(.services-table) {
-            width: 100% !important;
-            padding: 0 !important;
-            margin: 0 !important;
-            margin-right: 0 !important;
-          }
-          
-          /* Override any inline styles on table container */
-          [style*="marginRight"] {
-            margin-right: 0 !important;
-          }
-          table {
-            border-collapse: collapse !important;
-          }
-          
-          /* Ensure proper spacing */
-          .header {
-            margin-bottom: 30px !important;
-          }
-          
-          .details-grid {
-            margin-bottom: 30px !important;
-          }
-          
-          .total-section {
-            margin-top: 30px !important;
           }
         }
       `}</style>
